Store passwords in an encrypted and signed object
This commit adds some confidentiality and integrity assurances.

Confidentiality is obtained by encrypting the password before sending it
to the server. The domain and username are also somewhat protected since
they are hashed, however the hash is not keyed, so it is easy to attack.

Integrity is assured by a signature that covers not only the password
but also the username and domain.